# server_side/application/static/swagger.yaml
openapi: 3.0.0
info:
  title: Masters_Swagger
  contact: {}
  version: "1.0"
servers:
  - url: http:/127.0.0.1:5000
    variables: {}
paths:
...
  /login:
    post:
      tags:
        - User
      summary: Login
      operationId: Login
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              login: filip321
              password: pass
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}:
    get:
      tags:
        - User
      summary: GetUser
      operationId: GetUser
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 0
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false

    put:
      tags:
        - User
      summary: UpdateUser
      operationId: UpdateUser
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 1
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              login: filipk
              password: password
              email: email
              profile_photo:
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false

    delete:
      tags:
        - User
      summary: DeleteUser
      operationId: DeleteUser
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 1
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteUserRequest"
            example:
              login: filip
              password: haslo
              email: fifi
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
...
  /user:
    get:
      tags:
        - User
      summary: GetByLogin
      operationId: GetByLogin
      parameters:
        - name: login
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: filip
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/:
    put:
      tags:
        - User
      summary: UpdateSelf
      operationId: UpdateSelf
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSelfRequest"
            example:
              login: filip7
              password: password2
              email: filip_email32
              profile_photo:
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false

    delete:
      tags:
        - User
      summary: DeleteSelf
      operationId: DeleteSelf
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteSelfRequest"
            example:
              login: filip
              password: haslo
              email: fifi
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}/follow:
    put:
      tags:
        - User
      summary: FollowUser
      operationId: FollowUser
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 1
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUserRequest"
            example:
              login: filipk
              password: password
              email: email
              profile_photo:
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}/unfollow:
    put:
      tags:
        - User
      summary: UnFollowUser
      operationId: UnFollowUser
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 6
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnFollowUserRequest"
            example:
              login: filipk
              password: password
              email: email
              profile_photo:
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}/followed:
    get:
      tags:
        - User
      summary: GetFollowed
      operationId: GetFollowed
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 2
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}/followers:
    get:
      tags:
        - User
      summary: GetFollowers
      operationId: GetFollowers
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 6
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/followed:
    get:
      tags:
        - User
      summary: GetFollowedSelf
      operationId: GetFollowedSelf
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/followers:
    get:
      tags:
        - User
      summary: GetFollowersSelf
      operationId: GetFollowersSelf
      parameters:
        - name: limit
          in: query
          description: ""
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    LoginRequest:
      title: LoginRequest
      required:
        - login
        - password
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      example:
        login: filip321
        password: pass
    UpdateSelfRequest:
      title: UpdateSelfRequest
      required:
        - login
        - password
        - email
        - profile_photo
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
        profile_photo:
          type: string
          nullable: true
      example:
        login: filip7
        password: password2
        email: filip_email32
        profile_photo:
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
        - login
        - password
        - email
        - profile_photo
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
        profile_photo:
          type: string
          nullable: true
      example:
        login: filipk
        password: password
        email: email
        profile_photo:
    DeleteUserRequest:
      title: DeleteUserRequest
      required:
        - login
        - password
        - email
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
      example:
        login: filip
        password: haslo
        email: fifi
    DeleteSelfRequest:
      title: DeleteSelfRequest
      required:
        - login
        - password
        - email
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
      example:
        login: filip
        password: haslo
        email: fifi
    FollowUserRequest:
      title: FollowUserRequest
      required:
        - login
        - password
        - email
        - profile_photo
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
        profile_photo:
          type: string
          nullable: true
      example:
        login: filipk
        password: password
        email: email
        profile_photo:
    UnFollowUserRequest:
      title: UnFollowUserRequest
      required:
        - login
        - password
        - email
        - profile_photo
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
        profile_photo:
          type: string
          nullable: true
      example:
        login: filipk
        password: password
        email: email
        profile_photo:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...